Add new package to define general cache behavior and a simple Map based implementation

+++ trunk/uncertain/uncertain/cache/ICacheWriter.java
@@ -8,7 +8,7 @@ public interface ICacheWriter {
     
     public boolean setValue( Object key, Object value );
     
-    public boolean setValue( Object key, long timeout, Object value );
+    public boolean setValue( Object key, int timeout, Object value );
     
     public void remove( Object key );
